+++ appPlugins/ToolMilling.py
@@ -3260,8 +3260,10 @@ class ToolMilling(AppTool, Excellon):
 
             new_cncjob_obj.tools.update(tools_dict)
 
+            used_tools = list(tools_dict.keys())
+            new_cncjob_obj.used_tools = used_tools
             total_gcode = ''
-            for tooluid_key in list(tools_dict.keys()):
+            for tooluid_key in used_tools:
                 tool_cnt += 1
                 dia_cnc_dict = deepcopy(tools_dict[tooluid_key])